pull signature signing types into their own crate
