+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "mock",
     "rusoto/core",
     "rusoto/credential",
+    "rusoto/signature",
     "rusoto/services/acm",
     "rusoto/services/acm-pca",
     "rusoto/services/alexaforbusiness",
