+++ rusoto/signature/Cargo.toml
@@ -5,17 +5,15 @@ authors = [
     "Matthew Mayer <matthewkmayer@gmail.com>",
     "Nikita Pekin <contact@nikitapek.in>"
 ]
-build = "build.rs"
-description = "AWS SDK for Rust - Core"
-documentation = "https://docs.rs/rusoto_core"
+description = "AWS SDK for Rust - Request Signing"
+documentation = "https://docs.rs/rusoto_signature"
 keywords = ["AWS", "Amazon"]
 license = "MIT"
-name = "rusoto_core"
+name = "rusoto_signature"
 readme = "README.md"
 repository = "https://github.com/rusoto/rusoto"
 version = "0.41.0"
 homepage = "https://www.rusoto.org/"
-exclude = ["test_resources/*"]
 edition = "2018"
 
 [badges]
@@ -31,8 +29,6 @@ futures = "0.1.16"
 hmac = "0.7.1"
 http = "0.1.17"
 hyper = "0.12"
-hyper-tls = { version = "0.3.0", optional = true }
-hyper-rustls = { version = "0.17.1", optional = true }
 lazy_static = "1.0"
 log = "0.4.1"
 md5 = "0.6"
@@ -52,23 +48,9 @@ xml-rs = "0.8"
 path = "../credential"
 version = "0.41"
 
-[dependencies.rusoto_signature]
-path = "../signature"
-version = "0.41"
-
-[dependencies.clippy]
-optional = true
-version = "0.0"
-
 [dev-dependencies]
 env_logger = "0.5"
 rand = "0.4.2"
 serde_json = "1.0.1"
 serde_test = "1.0.1"
 
-[features]
-default = ["native-tls"]
-nightly-testing = ["rusoto_credential/nightly-testing"]
-native-tls = ["hyper-tls"]
-rustls = ["hyper-rustls"]
-unstable = []
