+++ rusoto/signature/src/lib.rs
@@ -1,0 +1,19 @@
+//! AWS API request signatures.
+//!
+//! Follows [AWS Signature 4](http://docs.aws.amazon.com/general/latest/gr/signature-version-4.html)
+//! algorithm.
+//!
+//! If needed, the request will be re-issued to a temporary redirect endpoint.  This can happen with
+//! newly created S3 buckets not in us-standard/us-east-1.
+//!
+//! Please note that this module does not expect URIs to already be encoded.
+//!
+#![cfg_attr(not(feature = "unstable"), deny(warnings))]
+pub extern crate rusoto_credential as credential;
+pub mod payload;
+pub mod region;
+pub mod signature;
+pub mod stream;
+pub use region::Region;
+pub use signature::{SignedRequest, SignedRequestPayload};
+pub use stream::ByteStream;
